instructional-python-planning/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
drf-spectacular==0.27.2